Refactor key vault innter resource to use a common abstract class
